+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Development SDK automatically sharing the built applet over proxy server called Forward Server Bridge (so the device doesn't have to be in the same network as the development machine)
+
 ## [1.25.0] - 2025-02-16
 ### Added
 - Extend applet commands in Timing for full JS API support for testing